refactor: pyproject.toml configuration
