+++ pyproject.toml
@@ -1,0 +1,26 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "sfetch_test_mihasket"
+version = "1.1.0"
+authors = [
+  { name="Miha Šket", email="sket.miha@proton.me" },
+]
+description = "Command-line tool for displaying spotify artist and album information. "
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dependencies = ["spotipy", "climage", "python-dotenv"]
+
+[project.urls]
+Homepage = "https://github.com/mihasket/sfetch"
+Issues = "https://github.com/mihasket/sfetch/issues"
+
+[project.scripts]
+sfetch = "sfetch.main:main"
